Changes through 20151029

+++ integration-tests/src/test/java/org/ihtsdo/otf/mapping/test/other/WorkflowActionNonLegacyPathTest.java
@@ -99,7 +99,11 @@ public class WorkflowActionNonLegacyPathTest {
       mappingService.removeMapRecord(mp.getId());
 
     for (MapUser mu : mappingService.getMapUsers().getIterable())
-      mappingService.removeMapUser(mu.getId());
+      if (!mu.getUserName().equals("guest")
+          && !mu.getUserName().equals("loader")
+          && !mu.getUserName().equals("qa")) {
+        mappingService.removeMapUser(mu.getId());
+      }
 
     for (TrackingRecord tr : workflowService.getTrackingRecords().getIterable())
       workflowService.removeTrackingRecord(tr.getId());
